+++ components/dialogs/transfer/components/max-balances-warning.tsx
@@ -1,6 +1,6 @@
 import { AlertTriangle } from "lucide-react"
 
-import { useIsMaxBalances } from "@/components/dialogs/transfer/use-is-max-balances"
+import { useIsMaxBalances } from "@/components/dialogs/transfer/hooks/use-is-max-balances"
 import {
   ResponsiveTooltip,
   ResponsiveTooltipContent,
