+++ components/dialogs/transfer/forms/default-form.tsx
@@ -12,21 +12,21 @@ import { mainnet } from "viem/chains"
 import { z } from "zod"
 
 import { TokenSelect } from "@/components/dialogs/token-select"
-import { BridgePrompt } from "@/components/dialogs/transfer/bridge-prompt"
+import { BridgePrompt } from "@/components/dialogs/transfer/components/bridge-prompt"
+import { MaxBalancesWarning } from "@/components/dialogs/transfer/components/max-balances-warning"
 import {
   ExternalTransferDirection,
-  checkAmount,
-  checkBalance,
+  isValidTransferAmount,
+  isBalanceSufficient,
   constructArbitrumBridgeUrl,
   formSchema,
   isMaxBalance,
 } from "@/components/dialogs/transfer/helpers"
 import { useChainBalance } from "@/components/dialogs/transfer/hooks/use-chain-balance"
+import { useIsMaxBalances } from "@/components/dialogs/transfer/hooks/use-is-max-balances"
 import { useRenegadeBalance } from "@/components/dialogs/transfer/hooks/use-renegade-balance"
 import { useToken } from "@/components/dialogs/transfer/hooks/use-token"
-import { MaxBalancesWarning } from "@/components/dialogs/transfer/max-balances-warning"
 import { Execution, Step, getSteps } from "@/components/dialogs/transfer/step"
-import { useIsMaxBalances } from "@/components/dialogs/transfer/use-is-max-balances"
 import { NumberInput } from "@/components/number-input"
 import { TokenIcon } from "@/components/token-icon"
 import { Button } from "@/components/ui/button"
@@ -220,22 +220,21 @@ export function DefaultForm({
   }
 
   async function onSubmit(values: z.infer<typeof formSchema>) {
-    const isAmountSufficient = checkAmount(queryClient, values.amount, l2Token)
-
     if (isDeposit) {
-      if (!isAmountSufficient) {
+      if (!isValidTransferAmount(queryClient, values.amount, l2Token)) {
         form.setError("amount", {
           message: `Amount must be greater than or equal to ${MIN_DEPOSIT_AMOUNT} USDC`,
         })
         return
       }
       await checkChain()
-      const isBalanceSufficient = checkBalance({
-        amount: values.amount,
-        mint: values.mint,
-        balance: l2Balance,
-      })
-      if (!isBalanceSufficient) {
+      if (
+        !isBalanceSufficient({
+          amount: values.amount,
+          mint: values.mint,
+          balance: l2Balance,
+        })
+      ) {
         form.setError("amount", {
           message: "Insufficient Arbitrum balance",
         })
@@ -271,7 +270,7 @@ export function DefaultForm({
     } else {
       // User is allowed to withdraw whole balance even if amount is < MIN_TRANSFER_AMOUNT
       if (
-        !isAmountSufficient &&
+        !isValidTransferAmount(queryClient, values.amount, l2Token) &&
         !isMaxBalance({
           amount: values.amount,
           mint: values.mint,
@@ -283,12 +282,13 @@ export function DefaultForm({
         })
         return
       }
-      const isBalanceSufficient = checkBalance({
-        amount: values.amount,
-        mint: values.mint,
-        balance: renegadeBalance,
-      })
-      if (!isBalanceSufficient) {
+      if (
+        !isBalanceSufficient({
+          amount: values.amount,
+          mint: values.mint,
+          balance: renegadeBalance,
+        })
+      ) {
         form.setError("amount", {
           message: "Insufficient Renegade balance",
         })
